apps/otto/pyproject.toml
[project]
name = "otto"
version = "0.1.0"
description = "Persistent AI agent for personal automation"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.scripts]
otto = "otto.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

